+++ asset_allocation/strategies/relative_value.py
@@ -114,11 +114,13 @@ class RelativeValue(strategy.Strategy):
             & (self.index_comp > 0)
             & (self.market_caps > self.market_cap_threshold)
             & (self.divyield > self.div_yield_threshold)
-            & (self.roe > self.roe_threshold)
+            # & (self.roe > self.roe_threshold)
             & (self.fcfps > self.freecashflow_threshold)
-            & (self.pe < self.spx_pe)
-            & (self.pb < self.spx_pb)
-            & (self.ps < self.spx_ps)
+            & (
+                (self.pe < self.spx_pe)
+                | (self.pb < self.spx_pb)
+                | (self.ps < self.spx_ps)
+            )
         ]
 
     @property