+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 ### Changed
 - run scaling in HRP Optimizer
+- Relative Value Strategy with OR instead of AND constraints for market multiples
 ### Removed
 
 ## [1.5.2] - 2023-12-26